requirements.txt
crewai
crewai_tools
streamlit 
pillow
streamlit
pytorch
langchain_cohere